pyproject.toml
[tool.poetry]
name = "hhporcelain"
version = "0.1.0"
description = "hannah ceramics repo"
authors = ["Benjamin Labaschin <benjaminlabaschin@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "3.10.1"
pandas = "^1.4.2"
jupyter = "^1.0.0"
ipykernel = "^6.13.0"
gspread-pandas = "^3.2.0"
streamlit = "^1.8.1"
google-auth = "^2.6.6"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
